pyproject.toml
[project]
name = "made-in-canada"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agno>=2.3.19",
    "bs4>=0.0.2",
    "cohere>=5.20.1",
    "coloredlogs>=15.0.1",
    "colorlog>=6.10.1",
    "httpx>=0.28.1",
    "openai>=2.14.0",
    "playwright>=1.57.0",
    "psycopg[binary]>=3.3.2",
    "python-decouple>=3.8",
    "sqlalchemy>=2.0.45",
    "streamlit>=1.52.2",
    "tqdm>=4.67.1",
]

[dependency-groups]
dev = [
    "ipdb>=0.13.13",
]
